requirements.txt
# Core dependencies
streamlit>=1.28.0
duckdb>=0.9.0
pandas>=2.1.0
requests>=2.31.0
python-dotenv>=1.0.0

# Data processing
openpyxl>=3.1.0
pyarrow>=14.0.0

# LLM integration
google-generativeai>=0.3.0

# Utilities
tenacity>=8.2.0
pydantic>=2.5.0
